Build: make build error fix
This patch is for fixing build error like below.

sometimes the build error occur which
obj/startup.d cannot be found.
The cause is that the obj folder cannot be found by
compiling without performing mkdir first in
the multi-thread environment like -j16.
so the code was modified to execute mkdir first.

Signed-off-by: chungho.shin <chungho.shin@nexell.co.kr>
Change-Id: I3f09462cddbdef67a8f0e40b0300f844b5ce9fe0